refactor: Optimize theme export and add docs
