+++ components/button/button-group.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames';
 import * as React from 'react';
 import { ConfigContext } from '../config-provider';
 import type { SizeType } from '../config-provider/SizeContext';
-import { useToken } from '../theme';
+import { useToken } from '../theme/internal';
 import warning from '../_util/warning';
 
 export interface ButtonGroupProps {
