+++ components/message/style/index.tsx
@@ -1,7 +1,7 @@
 // deps-lint-skip-all
 import { Keyframes } from '@ant-design/cssinjs';
-import type { FullToken, GenerateStyle } from '../../theme';
-import { genComponentStyleHook, mergeToken } from '../../theme';
+import type { FullToken, GenerateStyle } from '../../theme/internal';
+import { genComponentStyleHook, mergeToken } from '../../theme/internal';
 import { resetComponent } from '../../style';
 
 /** Component only token. Which will handle additional calculation of alias token */
