+++ components/theme/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
-import { defaultConfig, useToken as useInternalToken } from '.';
+import { defaultConfig, useToken as useInternalToken } from './internal';
 import defaultAlgorithm from './themes/default';
 import darkAlgorithm from './themes/dark';
 import compactAlgorithm from './themes/compact';
@@ -7,7 +7,7 @@ import compactAlgorithm from './themes/compact';
 // ZombieJ: We export as object to user but array in internal.
 // This is used to minimize the bundle size for antd package but safe to refactor as object also.
 // Please do not export internal `useToken` directly to avoid something export unexpected.
-/** Get current context Design Token. Will be different if you using nest theme config. */
+/** Get current context Design Token. Will be different if you are using nest theme config. */
 function useToken() {
   const [theme, token, hashId] = useInternalToken();
