+++ components/transfer/style/index.tsx
@@ -1,7 +1,7 @@
 import type { CSSObject } from '@ant-design/cssinjs';
 
-import type { FullToken, GenerateStyle } from '../../theme';
-import { genComponentStyleHook, mergeToken } from '../../theme';
+import type { FullToken, GenerateStyle } from '../../theme/internal';
+import { genComponentStyleHook, mergeToken } from '../../theme/internal';
 import { resetComponent, resetIcon, textEllipsis } from '../../style';
 
 export interface ComponentToken {
