+++ tests/setupAfterEnv.ts
@@ -2,7 +2,7 @@ import { toHaveNoViolations } from 'jest-axe';
 import '@testing-library/jest-dom';
 import format, { plugins } from 'pretty-format';
 import jsdom from 'jsdom';
-import { defaultConfig } from '../components/theme/index';
+import { defaultConfig } from '../components/theme';
 
 // Not use dynamic hashed for test env since version will change hash dynamically.
 defaultConfig.hashed = false;
@@ -15,8 +15,8 @@ if (process.env.LIB_DIR === 'dist') {
     return antd;
   });
 } else if (process.env.LIB_DIR === 'es') {
-  jest.mock('../es/theme', () => {
-    const esTheme = jest.requireActual('../es/theme');
+  jest.mock('../es/theme/internal', () => {
+    const esTheme = jest.requireActual('../es/theme/internal');
     if (esTheme.defaultConfig) {
       esTheme.defaultConfig.hashed = false;
     }
